flask_api/requirements.txt
numpy
urllib3
Pillow
flask
gunicorn
tensorflow
scikit-learn
pandas
joblib
flask_cors